Adapt calculators tf

Former-commit-id: 0274093b0c16bb4d10abdaf3a9f6fcbb24ca80bd

+++ tutorials/walkthrough_notebooks/conductivity_thermal/from_GK/conductivity_gk.py
@@ -6,11 +6,11 @@ import mdsuite as mds # Cool name for professionalism purposes
 
 new_case = True
 
-try:
-    new_case = True
-    shutil.rmtree('Argon')
-except FileNotFoundError:
-    pass
+if new_case:
+    try:
+        shutil.rmtree('Argon')
+    except FileNotFoundError:
+        pass
 
 
 
@@ -21,7 +21,7 @@ if new_case:
 
 # argon.help_computations_args('EinsteinDiffusionCoefficients') # auxiliary function to help on the extra arguments
 # argon.run_computation('EinsteinDiffusionCoefficients', species=list(argon.species.keys()),data_range=50, plot=False, singular=True, distinct=False)
-argon.run_computation('GreenKuboThermalConductivity', data_range=8000, plot=False)
+argon.run_computation('GreenKuboThermalConductivity', data_range=8000, plot=True, correlation_time=1)
 # argon.help_computations_args('GreenKuboDiffusionCoefficients')
-print(argon.results)
-argon.dump_results_json()
+print(type(argon.results))
+# argon.dump_results_json()
